requirements.txt
python-telegram-bot==20.7
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
beautifulsoup4==4.12.3
lxml==5.2.1
playwright==1.46.0
markdown==3.6
weasyprint==62.3
jinja2==3.1.4
matplotlib==3.8.4
mplfinance==0.12.10b0
pytz==2024.1
APScheduler==3.10.4
python-dotenv==1.0.1
